+++ core/classes/tools/class.tool.console.php
@@ -267,4 +267,10 @@ class ToolConsole
         global $neardLang, $neardTools;
         return $neardLang->getValue(Lang::COMPOSER) . ' ' . $neardTools->getComposer()->getVersion();
     }
+    
+    public function getTabTitlePhpUnit()
+    {
+        global $neardLang, $neardTools;
+        return $neardLang->getValue(Lang::PHPUNIT) . ' ' . $neardTools->getPhpUnit()->getVersion();
+    }
 }